# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/cases": {
      "get": {
        "operationId": "CasesController_findAll",
        "summary": "List GBV cases (county-scoped per user)",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cases"
        ]
      },
      "post": {
        "operationId": "CasesController_create",
        "summary": "Open a new GBV case — enters PENDING_REVIEW automatically",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
...
    "/v1/cases/{id}": {
      "get": {
        "operationId": "CasesController_findOne",
        "summary": "Get full case detail (audited)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/services": {
      "post": {
        "operationId": "CasesController_addService",
        "summary": "Record a service delivered on this case (blocked while in review)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/referrals": {
      "post": {
        "operationId": "CasesController_createReferral",
        "summary": "Create a referral for this case (blocked while in review)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/approve": {
      "post": {
        "operationId": "CasesController_approve",
        "summary": "Supervisor approves a PENDING_REVIEW case → OPEN",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/return": {
      "post": {
        "operationId": "CasesController_returnForRevision",
        "summary": "Supervisor returns a PENDING_REVIEW case for revision (notes required)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/resubmit": {
      "post": {
        "operationId": "CasesController_resubmit",
        "summary": "Case worker resubmits a RETURNED_FOR_REVISION case for review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
    "/v1/cases/{id}/close": {
      "post": {
        "operationId": "CasesController_close",
        "summary": "Close a case with reason",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cases"
        ]
      }
    },
...
  },
  "info": {
    "title": "GB MIS API",
    "description": "Gender-Based MIS — MOGCSP / LWEP",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}